% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{calc_pairwise_stability}
\alias{calc_pairwise_stability}
\title{Calculate stability for the importance results}
\usage{
calc_pairwise_stability(x, method, num.features = NULL)
}
\arguments{
\item{x}{list, data.frame or matrix of values to be tested for similarity}

\item{method}{one of \code{'kuncheva'}, \code{'sorensen'}, \code{'ochiai'}, \code{'jaccard'}}
}
\value{
numeric vector of similarities
}
\description{
Calculate stability for the importance results
}
