% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{check_fs_config}
\alias{check_fs_config}
\title{Check configuration of feature selection}
\usage{
check_fs_config(fsMethod, config, data)
}
\arguments{
\item{fsMethod}{feature selection method as specified in config file}

\item{config}{configuration list}

\item{data}{data frame to determine the number of features and samples}
}
\value{
list containing:
\item{fs_method}{feature selection method, see \link{link to feature selection}}
\item{fs_config}{configuration of feature selection}
}
\description{
Check configuration of feature selection
}
