% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{check_method_config}
\alias{check_method_config}
\title{Read method parameters}
\usage{
check_method_config(method_name, config)
}
\arguments{
\item{config}{list of configuration parameters}

\item{method}{string with the method name}
}
\value{
list with:
\item{method}{method, turned into \code{list}}
\item{method_config}{\code{list} of method parameters}
}
\description{
Read method parameters
}
