% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{check_optimization_config}
\alias{check_optimization_config}
\title{check optimization configuration}
\usage{
check_optimization_config(config, data, response.column = ".outcome")
}
\arguments{
\item{config}{list of configuration parameters}

\item{data}{data.frame with data}

\item{response.column}{string with the name of \code{response.column} to
determine the type of response and correspondingly the method valid for
optimization of parameters}
}
\value{
\code{NULL} if no optimization is anticipated and list otherwise,
containing:
\item{grid_resolution}{default value is \code{3}, unless specified in config}
\item{optimization_criterion}{\code{"performance"}, other accepted values are
\code{stability}, \code{RPT} as in \link{RPT paper}}
\item{metric}{for \code{performance}, depending on type of response, for regression it is
\code{"Rsquared_spearman"}, for classification it is \code{"AUC"}}
\item{internal_resampling}{default is "CV", can be any, see the \code{resamplingConfig}}
\item{internal_resamp_config}{configuration of internal resampling,
the same way as for external resampling}
}
\description{
For some methods, such as \code{glmnet} optimization is always required,
for other such as \code{rf} parameters can be used once and for all,
this is why for some methods we need to list methods that require optimization
vs those that do not
}
