% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{check_resamp_config}
\alias{check_resamp_config}
\title{Check configuration of resampling}
\usage{
check_resamp_config(config)
}
\arguments{
\item{config}{list of configuration parameters}
}
\value{
list of resampling configuration parameters containing:
\item{resampMethod}{method of resampling, such as \code{'CV'}, \code{'bootstrap'}}
}
\description{
Check configuration of resampling
}
