% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{configurePipeline}
\alias{configurePipeline}
\title{configure the training}
\usage{
configurePipeline(configFile, data, response.column)
}
\arguments{
\item{configFile}{name of the file with configuration}

\item{data}{data.frame with features and response}
}
\value{
list containing:
\item{fs_config}{feature selection configuration list}
\item{resamp_config}{resampling configuration list}
\item{optimization_config}{optimization configuration list}
}
\description{
configure the training
}
