% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_resampling.R
\name{create_resampling}
\alias{create_resampling}
\title{Create resampling according to configurations specified by user}
\usage{
create_resampling(resamplingConfig, times, response)
}
\arguments{
\item{resamplingConfig}{list of configurations for the resampling of class
\code{resamplingConfig}, with the following accepted resampling methods:
The resampling method: \code{"bootstrap"}, \code{"Fraction"},
\code{"CV"}, the latter as number of folds \code{k} accepts value \code{"LOO"},}

\item{times}{number of repetitions for the resampling}

\item{data}{data matrix}
}
\value{
\code{resampleIndex} list of size \code{times} with indices for each resample
}
\description{
Create resampling according to configurations specified by user
}
