% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{estim_stability}
\alias{estim_stability}
\title{Estimate the stability across resamples}
\usage{
estim_stability(importance, method = "cor_spearman", num.features = NULL)
}
\arguments{
\item{importance}{data.frame or list of feature importances}

\item{method}{}
}
\value{
list containing:
\item{stab_mean}{ mean of pairwise stabilities}
\item{stab_sd}{standard deviation of pairwise stabilities}
}
\description{
Estimate the stability across resamples
}
