% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{read_config}
\alias{read_config}
\title{Read and parse configuration file}
\usage{
read_config(fileName)
}
\arguments{
\item{fileName}{path to the file with configuration options}
}
\value{
list with configuration parameters
}
\description{
Read and parse configuration file
}
