% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read the data from two files with feature table and responses}
\usage{
read_data(data.file, response.column, response.file = NULL,
  match.column = NULL, verbose = F)
}
\arguments{
\item{data.file}{the name of the file with the data matrix. It may or may not contain the response}

\item{response.column}{the name of response column}

\item{response.file}{if not NULL, the response table must be in this file}

\item{match.column}{the column (columns?)}
}
\value{
the list containing:
\item{data}{the data frame with features + response column named \code{.outcome}}
\item{response.column}{the name of the response columnn}
}
\description{
Read the data from two files with feature table and responses
}
