% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_features.R
\name{select_features}
\alias{select_features}
\title{select features with the feature selection method}
\usage{
select_features(importance, fs.method, fs.config, model = NULL, data = NULL)
}
\arguments{
\item{data}{}
}
\description{
select features with the feature selection method
}
