% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.importance.R
\name{summarize_importance}
\alias{summarize_importance}
\title{Summarize importance for the case of optimizing importance across algorithm parameters (draft from old script)}
\usage{
summarize_importance(result)
}
\arguments{
\item{result}{}
}
\description{
Summarize importance for the case of optimizing importance across algorithm parameters (draft from old script)
}
