% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_perf.R
\name{summarize_perf}
\alias{summarize_perf}
\title{Summarize the performance across the resampling iterations}
\usage{
summarize_perf(data, response.column, performance, metric = NULL)
}
\arguments{
\item{metric}{}
}
\value{
list containing of
\item{perf_mean}{mean performance}
\item{perf_sd}{standard deviation of performance}
}
\description{
Summarize the performance across the resampling iterations
}
