% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainStability.R
\name{train_main}
\alias{train_main}
\title{Main function to train the model and get a (ranked) list of features}
\usage{
train_main(data, response = ".outcome", method = "rf", fs.method = "TopN",
  fs.config = ifelse(fs.method = "TopN", list(top.n = 100), NULL),
  repeats = 100, resamplingConfig = list(method = "Fraction", p = 0.9),
  optim.config = optim_config, parallel = F, ...)
}
\arguments{
\item{data}{an object where samples are in rows and features are in columns
and one column is the response. This can be a data.frame or data.table.
Other types not tested}

\item{response}{a string specifying the name of the column with the response}

\item{method}{a string specifying which model to use}

\item{fs.method}{a string specifying which feature selection method to use}

\item{fs.config}{a list specifying configurations of the feature selection method}

\item{resamplingConfig}{configuration of resampling}

\item{times}{number of repetitions for resampling}
}
\value{
list containing:
\item{result}{TO FIX}
\item{next item}{TO FIX}
}
\description{
Main function to train the model and get a (ranked) list of features
}
