% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainStability.R
\name{train_with_feature_extraction}
\alias{train_with_feature_extraction}
\title{Workhorse function to extract model performance and feature list}
\usage{
train_with_feature_extraction(data, inTrain, outTrain, method, fs.method,
  fs.config, response = ".outcome", optim.config = NULL, ...)
}
\arguments{
\item{data}{data frame or data table used to build the model}

\item{inTrain}{indices to be included in the model}

\item{outTrain}{indices of the hold-out samples}

\item{method}{machine learning method to be used to build the model}

\item{fs.method}{feature selection method to be used to select features}

\item{fs.config}{configurations of the feature selection model}

\item{response.column}{name of the column with responses}

\item{resampleIndex}{list of the integer indices of samples to be included in
the model}
}
\value{
A list is returned  containing: \item{method
}{the object of class \code{train}.see \code{\link[caret]{train}}}
\item{importance}{list of variables (predictors) with their corresponding
importance values or coefficients, see \link[caret]{varImp}}
\item{results }{a data frame the training error rate and values of the
tuning parameters.} \item{bestTune }{a data frame with the final
parameters.}
}
\description{
Workhorse function to extract model performance and feature list
}
\author{
Jelena Chuklina (the guts of \code{train.with.feature.extraction}
is a wrapper aroung the functions of \code{\link[caret]{train}})
}
